feat: add submissions table output

+++ backend/src/main/java/com/example/crazecode/domain/dto/SubmissionDto.java
@@ -9,11 +9,11 @@ import java.util.List;
 @Setter
 @Getter
 public class SubmissionDto implements Serializable {
-    private int language_id;
+    private Long language_id;
     private String source_code;
     private String expected_output;
 
-    public SubmissionDto(int language_id, String source_code, String expected_output) {
+    public SubmissionDto(Long language_id, String source_code, String expected_output) {
         this.language_id = language_id;
         this.source_code = source_code;
         this.expected_output = expected_output;
